+++ curionet/__init__.py
@@ -1,5 +1,5 @@
 from .model import CurioNet, CurioNetEncoder, CurioNetDecoder
-from .curiosity import CuriosityLayer, CuriosityBlock, WonderGenerator, WonderConv, InsightExtractor
+from .curiosity import CuriosityLayer, CuriosityBlock, WonderGenerator, WonderMixer, InsightExtractor
 from .transformer import TransformerSeq2Seq
 from .tokenizer import CharTokenizer
 from .data import WikiText2Dataset, make_vocab, VOCAB_SIZE